Convert PIP's `<Header>` to Typescript
